factory-config-vpn: check for connection state via nmconnection file
Depending on how fast the system performs the server and client
wireguard checks, it might end up getting the wrong result as the NM
connection removal is not done in a synchronous way (it can show up at
the nmcli conn output before fully removed).

Fix the race condition by just using the connection file instead when
checking if the vpn was previously set, as then we don't have to depend
on NM performing the full connection removal process.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>