vpn: Fix racey code path in initVpn
We have a scenario where a device could:

 * create the wg privkey
 * it's written to /var/sota/wg-priv.
 * the public key fails to upload
 * device powers down or fioconfig dies

When fioconfig resumes, it sees the file exists and won't upload it to
the server. This change writes to a temp file first and then commits the
change once we are successful.

This is safe because the server will always allow the device to update
this file.

Signed-off-by: Andy Doan <andy@foundries.io>