Introduce the framework for certificate rotation
Certificate rotation will work as a series of a few steps we must
execute in order. We must maintain a state where we can resume executing
these steps in the event of a power failure.

This change introduces data structure CertRotationState. As steps are
added to the rotation procedure, this struct will have new fields added
the are required for serialization.

There winds up being some boiler plate logic required to execute each
step (like saving state and handling errors). A generic
"CertRotationStep" interface was defined so that each unique step can
implement this interface.

Finally, a CertRotationHandler class was created to manage the execution
of these steps. Abstracting the steps also makes it much easier to write
test code to exercise the main handler logic w/o worrying about the side
effects of the real rotation steps.

Signed-off-by: Andy Doan <andy@foundries.io>