Cleanup: link unmodified internal ecies functions instead of copying
This is somewhat speculative change as it uses "unsafe" directive.
Ideally, I think we need to fork the umbrace/ecies package and add pkcs11 support to it.
However, this change have further cut the copy-pasted code to the bare minimum.
So, maybe we should keep it, or maybe we should remove this commit.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>