contrib: Simplify aktualizr config handler
We've tried to maintain this clever way of commenting out config items
managed by fioctl/fioconfig. However, its never been easy. With the
introduction of change:

 https://github.com/foundriesio/fioconfig/commit/7319b01c90b99847782f07f28734f90159146093

We no longer get the exact name of sota.toml. This change helps compute
the file name from the list of paths we get. If not found, we'll be
graceful.

Signed-off-by: Andy Doan <andy@foundries.io>