contrib: Add on-change handler for run-and-report
This provides a mechanism by which we can trigger remote actions using
fioconfig and the new `run-and-report` command. In order to prevent the
execution of random commands, we include a "actions" directory where a
user can be explicit about which remote actions they want to allow for
their product.

This commit includes a "reboot" command that's been requested in the
past.

Signed-off-by: Andy Doan <andy@foundries.io>