Add init logic to configure remote actions
This change helps the user facing UX part of this feature. By sharing
what actions are availble to the device, fioctl and webui can know what
options to present to the user for triggering a remote action. It can
also *reject* an invalid trigger rather than waiting for it to fail.