Define new type of init function callback
The current mechanism is really tied to fioconfig and doesn't make it
easy for us to add a new init function that also needs to send a config
file to the server.

It would be much easier to just let each init function make the call to
the server. However, doing this as a single request makes things a scale
slightly more resilient (one HTTP call instead of $X)

Signed-off-by: Andy Doan <andy@foundries.io>