transport: Avoid exception when printing warning after http error
httpDoOnce may return nil for the response in case of errors. This was
leading to a nil pointer dereference when trying to log the response.
This patch adds a check to avoid accessing the response status code in
that case.
A second check is also added on the break condition, when err != nil.
This should not be needed with the current httpDoOnce implementation,
but is added for safety in case of future changes.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>